feat: クリップ同期とトークン更新を background に移設し期限付きトークンに対応
- 同期 fetch を content script から background service worker へ移設。
  content の fetch はページオリジンの CORS でサイト API に 403 で弾かれ、
  Netflix/Disney からの同期が恒久失敗していた(サイト側 CLIP_API_ALLOWED_ORIGINS
  に chrome-extension://<ID> の登録が必要)
- サーバ新設の POST /api/extension/token/refresh に合わせ、chrome.alarms
  (6時間毎)+SW 起動時に残り15日未満でトークンをローテーション更新。
  JWT 前提の旧 checkAndRenewToken/decodeBase64Url/getTokenExpiryMs は削除
- pendingClips の再送 alarm(15分毎)と SW 起動時 flush を追加
- EXT_LINK_WITH_AUTH_TOKEN の expiresAt を保存、EXTENSION_UNLINKED 受信で
  トークン破棄(連携解除)。EXTENSION_AUTH_STATUS に linked を併記
- sync と refresh は runExclusive で直列化し、ローテーション中の旧トークン
  401 によるトークン誤クリアを防止
