+++ manifest.json
@@ -8,7 +8,8 @@
     "activeTab",
     "storage",
     "tabs",
-    "scripting" 
+    "scripting",
+    "alarms"
   ],
 
   "host_permissions": [
@@ -19,7 +20,7 @@
   ],
 
   "background": {
-    "service_worker": "src/background/background.js"
+    "service_worker": "dist/background.js"
   },
 
   "content_scripts": [
